+++ src/main/java/com/almostreliable/kubeoritech/event/BlackHoleEvent.java
@@ -1,4 +1,0 @@
-package com.almostreliable.kubeoritech.event;
-
-public class BlackHoleEvent {
-}